requirements.txt
pytesseract
Pillow
phonenumbers
extract-emails
phonenumbers
regex
